Make documents router dependent on app token
