+++ tests/non_search/routers/documents/test_get_document_families.py
@@ -1,9 +1,13 @@
 import pytest
 from db_client.models.dfce.family import Family, FamilyDocument, FamilyEvent
+from fastapi import status
 from fastapi.testclient import TestClient
 from sqlalchemy import update
 from sqlalchemy.orm import Session
 
+from tests.non_search.routers.documents.setup_doc_fam_lookup import (
+    _make_doc_fam_lookup_request,
+)
 from tests.non_search.setup_helpers import (
     setup_with_docs,
     setup_with_two_docs,
@@ -16,43 +20,41 @@ N_DOCUMENT_KEYS = 12
 
 
 def test_documents_family_slug_returns_not_found(
-    data_db: Session,
-    data_client: TestClient,
+    data_db: Session, data_client: TestClient, valid_token
 ):
     setup_with_docs(data_db)
     assert data_db.query(Family).count() == 1
     assert data_db.query(FamilyEvent).count() == 1
 
     # Test by slug
-    response = data_client.get(
-        "/api/v1/documents/FamSlug100",
+    json_response = _make_doc_fam_lookup_request(
+        data_client,
+        valid_token,
+        "FamSlug100",
+        expected_status_code=status.HTTP_404_NOT_FOUND,
     )
-    assert response.status_code == 404
-    assert response.json()["detail"] == "Nothing found for FamSlug100"
+    assert json_response["detail"] == "Nothing found for FamSlug100"
 
 
 def test_documents_family_slug_returns_correct_family(
-    data_db: Session,
-    data_client: TestClient,
+    data_db: Session, data_client: TestClient, valid_token
 ):
     setup_with_two_docs(data_db)
 
     # Test by slug
-    response = data_client.get(
-        "/api/v1/documents/FamSlug1",
+    json_response = _make_doc_fam_lookup_request(
+        data_client,
+        valid_token,
+        "FamSlug1",
     )
-
-    json_response = response.json()
-    assert response.status_code == 200
     assert json_response["import_id"] == "CCLW.family.1001.0"
 
     # Ensure a different family is returned
-    response = data_client.get(
-        "/api/v1/documents/FamSlug2",
+    json_response = _make_doc_fam_lookup_request(
+        data_client,
+        valid_token,
+        "FamSlug2",
     )
-
-    json_response = response.json()
-    assert response.status_code == 200
     assert json_response["import_id"] == "CCLW.family.2002.0"
 
 
@@ -124,17 +126,21 @@ def test_documents_family_slug_returns_correct_family(
     ],
 )
 def test_documents_family_slug_returns_correct_json(
-    data_client: TestClient, data_db: Session, slug, expected_fam, expected_doc
+    data_client: TestClient,
+    data_db: Session,
+    slug,
+    expected_fam,
+    expected_doc,
+    valid_token,
 ):
     setup_with_two_docs(data_db)
 
     # Test associations
-    response = data_client.get(
-        f"/api/v1/documents/{slug}",
+    json_response = _make_doc_fam_lookup_request(
+        data_client,
+        valid_token,
+        slug,
     )
-    json_response = response.json()
-
-    assert response.status_code == 200
 
     # Verify family data correct.
     assert len(json_response) == N_FAMILY_KEYS
@@ -170,23 +176,20 @@ def test_documents_family_slug_returns_correct_json(
 
 
 def test_documents_family_slug_returns_multiple_docs(
-    data_client: TestClient,
-    data_db: Session,
+    data_client: TestClient, data_db: Session, valid_token
 ):
     setup_with_two_docs_one_family(data_db)
 
-    response = data_client.get(
-        "/api/v1/documents/FamSlug1",
+    json_response = _make_doc_fam_lookup_request(
+        data_client,
+        valid_token,
+        "FamSlug1",
     )
-    json_response = response.json()
-
-    assert response.status_code == 200
     assert len(json_response["documents"]) == 2
 
 
 def test_documents_family_slug_returns_only_published_docs(
-    data_client: TestClient,
-    data_db: Session,
+    data_client: TestClient, data_db: Session, valid_token
 ):
     setup_with_two_docs_one_family(data_db)
     data_db.execute(
@@ -196,18 +199,16 @@ def test_documents_family_slug_returns_only_published_docs(
     )
 
     # Test associations
-    response = data_client.get(
-        "/api/v1/documents/FamSlug1",
+    json_response = _make_doc_fam_lookup_request(
+        data_client,
+        valid_token,
+        "FamSlug1",
     )
-    json_response = response.json()
-
-    assert response.status_code == 200
     assert len(json_response["documents"]) == 1
 
 
 def test_documents_family_slug_returns_404_when_all_docs_deleted(
-    data_client: TestClient,
-    data_db: Session,
+    data_client: TestClient, data_db: Session, valid_token
 ):
     setup_with_two_docs_one_family(data_db)
     data_db.execute(
@@ -222,29 +223,30 @@ def test_documents_family_slug_returns_404_when_all_docs_deleted(
     )
 
     # Test associations
-    response = data_client.get(
-        "/api/v1/documents/FamSlug1",
+    json_response = _make_doc_fam_lookup_request(
+        data_client,
+        valid_token,
+        "FamSlug1",
+        expected_status_code=status.HTTP_404_NOT_FOUND,
     )
-    json_response = response.json()
-
-    assert response.status_code == 404
     assert json_response["detail"] == "Family CCLW.family.1001.0 is not published"
 
 
 def test_documents_doc_slug_returns_not_found(
-    data_client: TestClient,
-    data_db: Session,
+    data_client: TestClient, data_db: Session, valid_token
 ):
     setup_with_docs(data_db)
     assert data_db.query(Family).count() == 1
     assert data_db.query(FamilyEvent).count() == 1
 
     # Test associations
-    response = data_client.get(
-        "/api/v1/documents/DocSlug100",
+    json_response = _make_doc_fam_lookup_request(
+        data_client,
+        valid_token,
+        "DocSlug100",
+        expected_status_code=status.HTTP_404_NOT_FOUND,
     )
-    assert response.status_code == 404
-    assert response.json()["detail"] == "Nothing found for DocSlug100"
+    assert json_response["detail"] == "Nothing found for DocSlug100"
 
 
 @pytest.mark.parametrize(
@@ -307,15 +309,20 @@ def test_documents_doc_slug_returns_not_found(
     ],
 )
 def test_documents_doc_slug_preexisting_objects(
-    data_client: TestClient, data_db: Session, slug, expected_fam, expected_doc
+    data_client: TestClient,
+    data_db: Session,
+    slug,
+    expected_fam,
+    expected_doc,
+    valid_token,
 ):
     setup_with_two_docs(data_db)
 
-    response = data_client.get(
-        f"/api/v1/documents/{slug}",
+    json_response = _make_doc_fam_lookup_request(
+        data_client,
+        valid_token,
+        slug,
     )
-    json_response = response.json()
-    assert response.status_code == 200
     assert len(json_response) == 2
 
     family = json_response["family"]
@@ -330,8 +337,7 @@ def test_documents_doc_slug_preexisting_objects(
 
 
 def test_documents_doc_slug_when_deleted(
-    data_client: TestClient,
-    data_db: Session,
+    data_client: TestClient, data_db: Session, valid_token
 ):
     setup_with_two_docs(data_db)
     data_db.execute(
@@ -339,9 +345,11 @@ def test_documents_doc_slug_when_deleted(
         .where(FamilyDocument.import_id == "CCLW.executive.2.2")
         .values(document_status="Deleted")
     )
-    response = data_client.get(
-        "/api/v1/documents/DocSlug2",
+
+    json_response = _make_doc_fam_lookup_request(
+        data_client,
+        valid_token,
+        "DocSlug2",
+        expected_status_code=status.HTTP_404_NOT_FOUND,
     )
-    json_response = response.json()
-    assert response.status_code == 404
     assert json_response["detail"] == "The document CCLW.executive.2.2 is not published"
